♻️ Refactor : 커피챗 filter 리팩토링
