+++ src/pages/coffeechats/FiltersAndButton.jsx
@@ -1,8 +1,9 @@
 import { Box, Button } from "@mui/material";
-import { Filters } from "components/common/filters/Filters";
+
 import { useNavigate } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import { isLoggedInState } from "recoil/atoms";
+import { Filters } from "./components/Filters";
 
 function FiltersAndButton({ sortData, kindOfJd, onChange }) {
   const navigate = useNavigate();
